+++ allennlp/data/tokenizers/letters_digits_tokenizer.py
@@ -1,7 +1,6 @@
 import re
 from typing import List
 
-from overrides import overrides
 
 from allennlp.data.tokenizers.token_class import Token
 from allennlp.data.tokenizers.tokenizer import Tokenizer
@@ -16,7 +15,6 @@ class LettersDigitsTokenizer(Tokenizer):
     Registered as a `Tokenizer` with name "letters_digits".
     """
 
-    @overrides
     def tokenize(self, text: str) -> List[Token]:
         # We use the [^\W\d_] pattern as a trick to match unicode letters
         tokens = [Token(m.group(), idx=m.start()) for m in re.finditer(r"[^\W\d_]+|\d+|\S", text)]