+++ allennlp/models/basic_classifier.py
@@ -1,6 +1,6 @@
 from typing import Dict, Optional
 
-from overrides import overrides
+
 import torch
 
 from allennlp.data import TextFieldTensors, Vocabulary
@@ -145,7 +145,6 @@ class BasicClassifier(Model):
 
         return output_dict
 
-    @overrides
     def make_output_human_readable(
         self, output_dict: Dict[str, torch.Tensor]
     ) -> Dict[str, torch.Tensor]: