+++ allennlp/modules/backbones/vilbert_backbone.py
@@ -2,7 +2,7 @@ import logging
 from typing import Dict, List
 
 import torch
-from overrides import overrides
+
 
 from allennlp.data.fields.text_field import TextFieldTensors
 from allennlp.data.vocabulary import Vocabulary
@@ -103,9 +103,8 @@ class VilbertBackbone(Backbone):
             fusion_method=fusion_method,
         )
 
-    @overrides
-    def forward(
-        self,  # type: ignore
+    def forward(  # type: ignore
+        self,
         box_features: torch.Tensor,
         box_coordinates: torch.Tensor,
         box_mask: torch.Tensor,