+++ allennlp/modules/elmo.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, List, Union
 
 import numpy
 import torch
-from overrides import overrides
+
 from torch.nn.modules import Dropout
 
 from allennlp.common import FromParams
@@ -313,7 +313,6 @@ class _ElmoCharacterEncoder(torch.nn.Module):
     def get_output_dim(self):
         return self.output_dim
 
-    @overrides
     def forward(self, inputs: torch.Tensor) -> Dict[str, torch.Tensor]:
         """
         Compute context insensitive token embeddings for ELMo representations.