+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed the docstring information for the `FBetaMultiLabelMeasure` metric.
 - Various fixes for Python 3.9
 
+### Removed
+
+- Removed dependency on the overrides package
+
+
 ## [v2.8.0](https://github.com/allenai/allennlp/releases/tag/v2.8.0) - 2021-11-01
 
 ### Added